app/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
httpx==0.25.2
openai==1.3.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
# Using html2text instead of readability-lxml to avoid lxml compilation issues
html2text==2020.1.16
# Alternative: trafilatura for text extraction (no lxml compilation needed)
trafilatura==1.6.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
ruff==0.1.6
mypy==1.7.0